Remove test in layer 4 'interface adapters'.
The problem is, that interface adapters are inherently bound to 3rd party code which can not be mocked.
This is in the nature of interface adapters and the reason why they are created in first place.

The inability to mock that layer makes it impossible to compile a unit test for native platform.