Remove unused sequence diagram.
This was used to brainstorm about whether Controllers and Presenters shall be part of
the 'interface adapters' or the 'board adapters' or both.

In the end I decided that in general they would be part of 'board adapters'
(which is a special kind of interface adapters).
But the software architecture does not need to make any requirements in that matter.